Rearrange files, refactor code
