+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "react-dev": "webpack -d --watch",
-    "server-dev": "nodemon server/index.js",
-    "seed": "node server/data/seedData.js",
+    "server-dev": "nodemon server/server.js",
+    "seed": "node server/database-mongo/data/seedData.js",
     "test": "jest"
   },
   "repository": {
