+++ server/controllers/app.js
@@ -2,14 +2,10 @@ const express = require('express');
 const path = require('path');
 
 const bodyParser = require('body-parser');
-
-const mongoose = require('mongoose');
-// eslint-disable-next-line node/no-unpublished-require
-const dbUrl = require('./database-mongo/database.config');
-const Product = require('./models/Product');
+// const db = require('../database-mongo/index');
+const Product = require('../models/Product');
 
 const app = express();
-const port = process.env.PORT || 8080;
 
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(express.json());
@@ -28,12 +24,4 @@ app.get('/youmayalsolike', (req, res) => {
     });
 });
 
-//  Start server and db
-mongoose
-  .connect(dbUrl, { useUnifiedTopology: true, useNewUrlParser: true })
-  .then(() => {
-    app.listen(port, () => console.log(`Listening on port: ${port}`));
-  })
-  .catch((err) => {
-    console.log(err);
-  });
+module.exports = app;
