+++ server/database-mongo/index.js
@@ -1,0 +1,8 @@
+// const mongoose = require('mongoose');
+// // eslint-disable-next-line node/no-unpublished-require
+// const dbUrl = require('./config/database.config');
+
+// module.exports = mongoose.connect(dbUrl, {
+//   useUnifiedTopology: true,
+//   useNewUrlParser: true
+// });
